+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [1.55.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.54.1...v1.55.0) (2025-02-19)
+
+
+### Features
+
+* add image alt-text audit ([#647](https://github.com/adobe/spacecat-audit-worker/issues/647)) ([3e9309a](https://github.com/adobe/spacecat-audit-worker/commit/3e9309a4c7550a8454624f163b68724d8a4bed97))
+
+## [1.54.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.54.0...v1.54.1) (2025-02-19)
+
+
+### Bug Fixes
+
+* step audit queue url ([#672](https://github.com/adobe/spacecat-audit-worker/issues/672)) ([15a0c33](https://github.com/adobe/spacecat-audit-worker/commit/15a0c339d60434154f89a3dfc8c9b4d61541462e))
+
+# [1.54.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.53.2...v1.54.0) (2025-02-18)
+
+
+### Features
+
+* audit steps ([#659](https://github.com/adobe/spacecat-audit-worker/issues/659)) ([8b0362d](https://github.com/adobe/spacecat-audit-worker/commit/8b0362dae4c57440d9f786d4dfe74105df2124ca))
+
+## [1.53.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.53.1...v1.53.2) (2025-02-18)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.6.0 ([#667](https://github.com/adobe/spacecat-audit-worker/issues/667)) ([4fbd3db](https://github.com/adobe/spacecat-audit-worker/commit/4fbd3db2481c849e175873737aeef2a019c9ee6f))
+
 ## [1.53.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.53.0...v1.53.1) (2025-02-18)